+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Implemented creation of payments through API redirecting to link, instead of posting form directly to onpay
+- Updated SDK version
 
 ## [1.0.25] - 2022-06-08
 - Added support for customers updating payment method